Refactor service module making it more stable
